refactor(sfn+sqs): split stepfunctions/service.rs and sqs/service.rs by family
Audit 2026-05-19 P9 continuation.

Stepfunctions (2820 -> 1489 mod + 7 submodules):
- state_machines.rs 566  executions.rs 435  activities.rs 139
- map_runs.rs 101  tags.rs 88  tasks.rs 64  validation.rs 25

SQS (2912 -> 502 mod + 7 submodules):
- messages (send/receive/delete/change-vis/encrypt+decrypt body)
- queues (create/delete/purge/url/list/attrs)
- tags / permissions / dlq / moves / auth helpers

Same pattern as PR #1479/#1480/...: pub(super) family methods, helpers
+ service_tests reach via #[path] attrs.

sfn 161 + sqs 148 unit tests pass.
