+++ crates/fakecloud-stepfunctions/src/service/mod.rs
@@ -78,6 +78,14 @@ pub struct StepFunctionsService {
     snapshot_lock: Arc<AsyncMutex<()>>,
 }
 
+mod activities;
+mod executions;
+mod map_runs;
+mod state_machines;
+mod tags;
+mod tasks;
+mod validation;
+
 impl StepFunctionsService {
     pub fn new(state: SharedStepFunctionsState) -> Self {
         Self {
@@ -228,668 +236,6 @@ impl AwsService for StepFunctionsService {
 }
 
 impl StepFunctionsService {
-    // ─── State Machine CRUD ─────────────────────────────────────────────
-
-    fn create_state_machine(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        validate_required("name", &body["name"])?;
-        let name = body["name"].as_str().ok_or_else(|| missing("name"))?;
-        validate_name(name)?;
-
-        validate_required("definition", &body["definition"])?;
-        let definition = body["definition"]
-            .as_str()
-            .ok_or_else(|| missing("definition"))?;
-        validate_definition(definition)?;
-
-        validate_required("roleArn", &body["roleArn"])?;
-        let role_arn = body["roleArn"].as_str().ok_or_else(|| missing("roleArn"))?;
-        validate_arn(role_arn)?;
-
-        let machine_type = if let Some(t) = body["type"].as_str() {
-            StateMachineType::parse(t).ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!(
-                        "Value '{t}' at 'type' failed to satisfy constraint: \
-                         Member must satisfy enum value set: [STANDARD, EXPRESS]"
-                    ),
-                )
-            })?
-        } else {
-            StateMachineType::Standard
-        };
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let arn = state.state_machine_arn(name);
-
-        // Check if name already exists
-        if state.state_machines.values().any(|sm| sm.name == name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::CONFLICT,
-                "StateMachineAlreadyExists",
-                format!("State Machine Already Exists: '{arn}'"),
-            ));
-        }
-
-        let now = Utc::now();
-        let revision_id = uuid::Uuid::new_v4().to_string();
-
-        let mut tags = BTreeMap::new();
-        if !body["tags"].is_null() {
-            fakecloud_core::tags::apply_tags(&mut tags, &body, "tags", "key", "value").map_err(
-                |f| {
-                    AwsServiceError::aws_error(
-                        StatusCode::BAD_REQUEST,
-                        "ValidationException",
-                        format!("{f} must be a list"),
-                    )
-                },
-            )?;
-        }
-
-        let sm = StateMachine {
-            name: name.to_string(),
-            arn: arn.clone(),
-            definition: definition.to_string(),
-            role_arn: role_arn.to_string(),
-            machine_type,
-            status: StateMachineStatus::Active,
-            creation_date: now,
-            update_date: now,
-            tags,
-            revision_id: revision_id.clone(),
-            logging_configuration: body.get("loggingConfiguration").cloned(),
-            tracing_configuration: body.get("tracingConfiguration").cloned(),
-            description: body["description"].as_str().unwrap_or("").to_string(),
-        };
-
-        state.state_machines.insert(arn.clone(), sm);
-
-        Ok(AwsResponse::ok_json(json!({
-            "stateMachineArn": arn,
-            "creationDate": now.timestamp() as f64,
-            "stateMachineVersionArn": arn,
-        })))
-    }
-
-    fn describe_state_machine(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("stateMachineArn", &body["stateMachineArn"])?;
-        let arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?;
-        validate_arn(arn)?;
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let sm = state
-            .state_machines
-            .get(arn)
-            .ok_or_else(|| state_machine_not_found(arn))?;
-
-        Ok(AwsResponse::ok_json(state_machine_to_json(sm)))
-    }
-
-    fn list_state_machines(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let raw_max_results = body["maxResults"].as_i64();
-        if let Some(mr) = raw_max_results {
-            validate_max_results(mr)?;
-        }
-        // Smithy default: maxResults=0 means "use the service default" (100).
-        let max_results = match raw_max_results.unwrap_or(0) {
-            0 => 100,
-            n => n as usize,
-        };
-        let next_token = body["nextToken"].as_str();
-        if let Some(t) = next_token {
-            validate_page_token(t)?;
-        }
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let mut machines: Vec<&StateMachine> = state.state_machines.values().collect();
-        machines.sort_by(|a, b| a.name.cmp(&b.name));
-
-        let items: Vec<Value> = machines
-            .iter()
-            .map(|sm| {
-                json!({
-                    "name": sm.name,
-                    "stateMachineArn": sm.arn,
-                    "type": sm.machine_type.as_str(),
-                    "creationDate": sm.creation_date.timestamp() as f64,
-                })
-            })
-            .collect();
-
-        let (page, token) = paginate(&items, next_token, max_results);
-
-        let mut resp = json!({ "stateMachines": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn delete_state_machine(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("stateMachineArn", &body["stateMachineArn"])?;
-        let arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?;
-        validate_arn(arn)?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        // AWS returns success even if it doesn't exist
-        state.state_machines.remove(arn);
-
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn update_state_machine(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("stateMachineArn", &body["stateMachineArn"])?;
-        let arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?;
-        validate_arn(arn)?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let sm = state
-            .state_machines
-            .get_mut(arn)
-            .ok_or_else(|| state_machine_not_found(arn))?;
-
-        if let Some(definition) = body["definition"].as_str() {
-            validate_definition(definition)?;
-            sm.definition = definition.to_string();
-        }
-
-        if let Some(role_arn) = body["roleArn"].as_str() {
-            validate_arn(role_arn)?;
-            sm.role_arn = role_arn.to_string();
-        }
-
-        if let Some(logging) = body.get("loggingConfiguration") {
-            sm.logging_configuration = Some(logging.clone());
-        }
-
-        if let Some(tracing) = body.get("tracingConfiguration") {
-            sm.tracing_configuration = Some(tracing.clone());
-        }
-
-        if let Some(description) = body["description"].as_str() {
-            sm.description = description.to_string();
-        }
-
-        let now = Utc::now();
-        sm.update_date = now;
-        sm.revision_id = uuid::Uuid::new_v4().to_string();
-
-        let revision_id = sm.revision_id.clone();
-        let sm_arn = sm.arn.clone();
-
-        Ok(AwsResponse::ok_json(json!({
-            "updateDate": now.timestamp() as f64,
-            "revisionId": revision_id,
-            "stateMachineVersionArn": sm_arn,
-        })))
-    }
-
-    // ─── Execution Lifecycle ──────────────────────────────────────────
-
-    fn start_execution(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("stateMachineArn", &body["stateMachineArn"])?;
-        let sm_arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?;
-        validate_arn(sm_arn)?;
-
-        let input = body["input"].as_str().map(|s| s.to_string());
-
-        // Validate input is valid JSON if provided
-        if let Some(ref input_str) = input {
-            let _: serde_json::Value = serde_json::from_str(input_str).map_err(|_| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidExecutionInput",
-                    "Invalid execution input: must be valid JSON".to_string(),
-                )
-            })?;
-        }
-
-        let execution_name = body["name"]
-            .as_str()
-            .map(|s| s.to_string())
-            .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-
-        if let Some(name) = body["name"].as_str() {
-            validate_name(name)?;
-        }
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let sm = state
-            .state_machines
-            .get(sm_arn)
-            .ok_or_else(|| state_machine_not_found(sm_arn))?;
-
-        let sm_name = sm.name.clone();
-        let definition = sm.definition.clone();
-        let exec_arn = state.execution_arn(&sm_name, &execution_name);
-
-        // Check for duplicate execution name
-        if state.executions.contains_key(&exec_arn) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::CONFLICT,
-                "ExecutionAlreadyExists",
-                format!("Execution Already Exists: '{exec_arn}'"),
-            ));
-        }
-
-        let now = Utc::now();
-        let execution = Execution {
-            execution_arn: exec_arn.clone(),
-            state_machine_arn: sm_arn.to_string(),
-            state_machine_name: sm_name,
-            name: execution_name,
-            status: ExecutionStatus::Running,
-            input: input.clone(),
-            output: None,
-            start_date: now,
-            stop_date: None,
-            error: None,
-            cause: None,
-            history_events: vec![],
-            parent_execution_arn: None,
-            is_sync: false,
-            billed_duration_ms: None,
-            billed_memory_mb: None,
-        };
-
-        state.executions.insert(exec_arn.clone(), execution);
-        let logging_config = sm.logging_configuration.clone();
-        drop(accounts);
-
-        // Spawn async execution
-        let shared_state = self.state.clone();
-        let exec_arn_clone = exec_arn.clone();
-        let input_clone = input;
-        let delivery = self.delivery.clone();
-        let dynamodb_state = self.dynamodb_state.clone();
-        let registry = self.registry.clone();
-        tokio::spawn(async move {
-            interpreter::execute_state_machine(
-                shared_state,
-                exec_arn_clone,
-                definition,
-                input_clone,
-                delivery,
-                dynamodb_state,
-                registry,
-                logging_config,
-            )
-            .await;
-        });
-
-        Ok(AwsResponse::ok_json(json!({
-            "executionArn": exec_arn,
-            "startDate": now.timestamp() as f64,
-        })))
-    }
-
-    fn stop_execution(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("executionArn", &body["executionArn"])?;
-        let exec_arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?;
-
-        let error = body["error"].as_str().map(|s| s.to_string());
-        let cause = body["cause"].as_str().map(|s| s.to_string());
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let exec = state
-            .executions
-            .get_mut(exec_arn)
-            .ok_or_else(|| execution_not_found(exec_arn))?;
-
-        if exec.status != ExecutionStatus::Running {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ExecutionNotRunning",
-                format!("Execution is not running: '{exec_arn}'"),
-            ));
-        }
-
-        let now = Utc::now();
-        exec.status = ExecutionStatus::Aborted;
-        exec.stop_date = Some(now);
-        exec.error = error;
-        exec.cause = cause;
-
-        Ok(AwsResponse::ok_json(json!({
-            "stopDate": now.timestamp() as f64,
-        })))
-    }
-
-    fn describe_execution(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("executionArn", &body["executionArn"])?;
-        let exec_arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?;
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let exec = state
-            .executions
-            .get(exec_arn)
-            .ok_or_else(|| execution_not_found(exec_arn))?;
-
-        Ok(AwsResponse::ok_json(execution_to_json(exec)))
-    }
-
-    fn list_executions(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("stateMachineArn", &body["stateMachineArn"])?;
-        let sm_arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?;
-        validate_arn(sm_arn)?;
-
-        let max_results = body["maxResults"].as_i64().unwrap_or(100) as usize;
-        validate_range_i64("maxResults", max_results as i64, 1, 1000)?;
-        let next_token = body["nextToken"].as_str();
-        let status_filter = body["statusFilter"].as_str();
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-
-        // Verify state machine exists
-        if !state.state_machines.contains_key(sm_arn) {
-            return Err(state_machine_not_found(sm_arn));
-        }
-
-        let mut executions: Vec<&Execution> = state
-            .executions
-            .values()
-            .filter(|e| e.state_machine_arn == sm_arn)
-            .filter(|e| {
-                status_filter
-                    .map(|sf| e.status.as_str() == sf)
-                    .unwrap_or(true)
-            })
-            .collect();
-
-        // Sort by start date descending (most recent first)
-        executions.sort_by_key(|e| std::cmp::Reverse(e.start_date));
-
-        let items: Vec<Value> = executions
-            .iter()
-            .map(|e| {
-                let mut item = json!({
-                    "executionArn": e.execution_arn,
-                    "stateMachineArn": e.state_machine_arn,
-                    "name": e.name,
-                    "status": e.status.as_str(),
-                    "startDate": e.start_date.timestamp() as f64,
-                });
-                if let Some(stop) = e.stop_date {
-                    item["stopDate"] = json!(stop.timestamp() as f64);
-                }
-                item
-            })
-            .collect();
-
-        let (page, token) = paginate(&items, next_token, max_results);
-
-        let mut resp = json!({ "executions": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn get_execution_history(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("executionArn", &body["executionArn"])?;
-        let exec_arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?;
-        validate_arn_length("executionArn", exec_arn, 256)?;
-
-        if let Some(mr) = body["maxResults"].as_i64() {
-            validate_max_results(mr)?;
-        }
-        let max_results = body["maxResults"].as_i64().unwrap_or(100) as usize;
-        let next_token = body["nextToken"].as_str();
-        if let Some(t) = next_token {
-            validate_page_token(t)?;
-        }
-        let reverse_order = body["reverseOrder"].as_bool().unwrap_or(false);
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let exec = state
-            .executions
-            .get(exec_arn)
-            .ok_or_else(|| execution_not_found(exec_arn))?;
-
-        let mut events: Vec<Value> = exec
-            .history_events
-            .iter()
-            .map(|e| {
-                json!({
-                    "id": e.id,
-                    "type": e.event_type,
-                    "timestamp": e.timestamp.timestamp() as f64,
-                    "previousEventId": e.previous_event_id,
-                    format!("{}EventDetails", camel_to_details_key(&e.event_type)): e.details,
-                })
-            })
-            .collect();
-
-        if reverse_order {
-            events.reverse();
-        }
-
-        let (page, token) = paginate(&events, next_token, max_results);
-
-        let mut resp = json!({ "events": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn describe_state_machine_for_execution(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("executionArn", &body["executionArn"])?;
-        let exec_arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?;
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let exec = state
-            .executions
-            .get(exec_arn)
-            .ok_or_else(|| execution_not_found(exec_arn))?;
-
-        let sm = state
-            .state_machines
-            .get(&exec.state_machine_arn)
-            .ok_or_else(|| state_machine_not_found(&exec.state_machine_arn))?;
-
-        Ok(AwsResponse::ok_json(state_machine_to_json(sm)))
-    }
-
-    // ─── Tagging ────────────────────────────────────────────────────────
-
-    fn tag_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("resourceArn", &body["resourceArn"])?;
-        let arn = body["resourceArn"]
-            .as_str()
-            .ok_or_else(|| missing("resourceArn"))?;
-        validate_arn(arn)?;
-        validate_required("tags", &body["tags"])?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let sm = state
-            .state_machines
-            .get_mut(arn)
-            .ok_or_else(|| resource_not_found(arn))?;
-
-        fakecloud_core::tags::apply_tags(&mut sm.tags, &body, "tags", "key", "value").map_err(
-            |f| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("{f} must be a list"),
-                )
-            },
-        )?;
-
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn untag_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("resourceArn", &body["resourceArn"])?;
-        let arn = body["resourceArn"]
-            .as_str()
-            .ok_or_else(|| missing("resourceArn"))?;
-        validate_arn(arn)?;
-        validate_required("tagKeys", &body["tagKeys"])?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let sm = state
-            .state_machines
-            .get_mut(arn)
-            .ok_or_else(|| resource_not_found(arn))?;
-
-        fakecloud_core::tags::remove_tags(&mut sm.tags, &body, "tagKeys").map_err(|f| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                format!("{f} must be a list"),
-            )
-        })?;
-
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn list_tags_for_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_required("resourceArn", &body["resourceArn"])?;
-        let arn = body["resourceArn"]
-            .as_str()
-            .ok_or_else(|| missing("resourceArn"))?;
-        validate_arn(arn)?;
-
-        let accounts = self.state.read();
-        let empty = StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let sm = state
-            .state_machines
-            .get(arn)
-            .ok_or_else(|| resource_not_found(arn))?;
-
-        let tags = fakecloud_core::tags::tags_to_json(&sm.tags, "key", "value");
-
-        Ok(AwsResponse::ok_json(json!({ "tags": tags })))
-    }
-
-    // ─── Activities ─────────────────────────────────────────────────────
-
-    fn create_activity(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let name = body["name"].as_str().ok_or_else(|| missing("name"))?;
-        validate_name(name)?;
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let arn = format!(
-            "arn:aws:states:{}:{}:activity:{}",
-            state.region, state.account_id, name
-        );
-        if state.activities.contains_key(&arn) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ActivityAlreadyExists",
-                format!("Activity already exists: {arn}"),
-            ));
-        }
-        let activity = crate::state::Activity {
-            name: name.to_string(),
-            arn: arn.clone(),
-            creation_date: chrono::Utc::now(),
-            tags: BTreeMap::new(),
-        };
-        state.activities.insert(arn.clone(), activity.clone());
-        Ok(AwsResponse::ok_json(json!({
-            "activityArn": arn,
-            "creationDate": activity.creation_date.timestamp(),
-        })))
-    }
-
-    fn delete_activity(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["activityArn"]
-            .as_str()
-            .ok_or_else(|| missing("activityArn"))?
-            .to_string();
-        validate_arn_length("activityArn", &arn, 256)?;
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        state.activities.remove(&arn);
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn describe_activity(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["activityArn"]
-            .as_str()
-            .ok_or_else(|| missing("activityArn"))?
-            .to_string();
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let a = state.activities.get(&arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ActivityDoesNotExist",
-                format!("Activity does not exist: {arn}"),
-            )
-        })?;
-        Ok(AwsResponse::ok_json(json!({
-            "activityArn": a.arn,
-            "name": a.name,
-            "creationDate": a.creation_date.timestamp(),
-        })))
-    }
-
     fn list_activities(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
         let body = req.json_body();
         let raw_max_results = body["maxResults"].as_i64();
@@ -928,684 +274,6 @@ impl StepFunctionsService {
             resp["nextToken"] = json!(t);
         }
         Ok(AwsResponse::ok_json(resp))
-    }
-
-    async fn get_activity_task(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["activityArn"]
-            .as_str()
-            .ok_or_else(|| missing("activityArn"))?
-            .to_string();
-        // Activity must exist before we'll accept long-poll calls.
-        {
-            let accounts = self.state.read();
-            let state = accounts
-                .get(&req.account_id)
-                .ok_or_else(|| activity_not_found(&arn))?;
-            if !state.activities.contains_key(&arn) {
-                return Err(activity_not_found(&arn));
-            }
-        }
-
-        // AWS GetActivityTask blocks up to 60s. fakecloud defaults to 5s
-        // so test suites don't stall when no worker is feeding the queue.
-        let max_wait_secs: u64 = std::env::var("FAKECLOUD_SFN_GET_ACTIVITY_TIMEOUT_SECS")
-            .ok()
-            .and_then(|s| s.parse().ok())
-            .unwrap_or(5);
-        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(max_wait_secs);
-
-        loop {
-            // Try to dequeue oldest PENDING token for this activity.
-            {
-                let mut accounts = self.state.write();
-                let state = accounts.get_or_create(&req.account_id);
-                let mut candidates: Vec<(String, chrono::DateTime<chrono::Utc>)> = state
-                    .task_tokens
-                    .iter()
-                    .filter(|(_, t)| t.activity_arn == arn && t.status == "PENDING")
-                    .map(|(k, t)| (k.clone(), t.created_at))
-                    .collect();
-                candidates.sort_by_key(|c| c.1);
-                if let Some((token, _)) = candidates.into_iter().next() {
-                    let now = chrono::Utc::now();
-                    let entry = state.task_tokens.get_mut(&token).expect("just looked up");
-                    entry.status = "IN_PROGRESS".to_string();
-                    entry.last_heartbeat_at = Some(now);
-                    let input = entry.input.clone().unwrap_or_else(|| "{}".to_string());
-                    return Ok(AwsResponse::ok_json(json!({
-                        "taskToken": token,
-                        "input": input,
-                    })));
-                }
-            }
-            if std::time::Instant::now() >= deadline {
-                // No task available in window — return empty token (matches
-                // AWS behavior).
-                return Ok(AwsResponse::ok_json(json!({
-                    "taskToken": "",
-                    "input": "",
-                })));
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-        }
-    }
-
-    fn send_task_success(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        self.update_task_token(req, "SUCCEEDED")
-    }
-
-    fn send_task_failure(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        self.update_task_token(req, "FAILED")
-    }
-
-    fn send_task_heartbeat(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        // Heartbeats only refresh `last_heartbeat_at`; they don't change
-        // the task's lifecycle status. The interpreter's heartbeat-timeout
-        // check reads `last_heartbeat_at` to decide whether to fail the
-        // task with `States.HeartbeatTimeout`.
-        let body = req.json_body();
-        let token = body["taskToken"]
-            .as_str()
-            .ok_or_else(|| missing("taskToken"))?
-            .to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let entry = state
-            .task_tokens
-            .get_mut(&token)
-            .ok_or_else(|| task_does_not_exist(&token))?;
-        entry.last_heartbeat_at = Some(chrono::Utc::now());
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn update_task_token(
-        &self,
-        req: &AwsRequest,
-        new_status: &str,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let token = body["taskToken"]
-            .as_str()
-            .ok_or_else(|| missing("taskToken"))?
-            .to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let entry = state
-            .task_tokens
-            .get_mut(&token)
-            .ok_or_else(|| task_does_not_exist(&token))?;
-        entry.status = new_status.to_string();
-        if new_status == "SUCCEEDED" {
-            entry.output = body["output"].as_str().map(String::from);
-        } else if new_status == "FAILED" {
-            entry.error = body["error"].as_str().map(String::from);
-            entry.cause = body["cause"].as_str().map(String::from);
-        }
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    // ─── State machine versions / aliases ───────────────────────────────
-
-    fn publish_state_machine_version(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?
-            .to_string();
-        let description = body["description"].as_str().unwrap_or("").to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        if !state.state_machines.contains_key(&arn) {
-            return Err(state_machine_not_found(&arn));
-        }
-        let version = state
-            .state_machine_versions
-            .values()
-            .filter(|v| v.state_machine_arn == arn)
-            .map(|v| v.version)
-            .max()
-            .unwrap_or(0)
-            + 1;
-        let version_arn = format!("{arn}:{version}");
-        let v = crate::state::StateMachineVersion {
-            state_machine_arn: arn,
-            version,
-            revision_id: format!("rev-{version}"),
-            description,
-            creation_date: chrono::Utc::now(),
-        };
-        state
-            .state_machine_versions
-            .insert(version_arn.clone(), v.clone());
-        Ok(AwsResponse::ok_json(json!({
-            "stateMachineVersionArn": version_arn,
-            "creationDate": v.creation_date.timestamp(),
-        })))
-    }
-
-    fn delete_state_machine_version(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineVersionArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineVersionArn"))?
-            .to_string();
-        validate_arn_length("stateMachineVersionArn", &arn, 2000)?;
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        state.state_machine_versions.remove(&arn);
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn list_state_machine_versions(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?
-            .to_string();
-        validate_arn_length("stateMachineArn", &arn, 256)?;
-        let raw_max_results = body["maxResults"].as_i64();
-        if let Some(mr) = raw_max_results {
-            validate_max_results(mr)?;
-        }
-        let next_token = body["nextToken"].as_str();
-        if let Some(t) = next_token {
-            validate_page_token(t)?;
-        }
-        let max_results = match raw_max_results.unwrap_or(0) {
-            0 => 100,
-            n => n as usize,
-        };
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let mut versions: Vec<&crate::state::StateMachineVersion> = state
-            .state_machine_versions
-            .values()
-            .filter(|v| v.state_machine_arn == arn)
-            .collect();
-        versions.sort_by_key(|v| std::cmp::Reverse(v.version));
-        let items: Vec<Value> = versions
-            .iter()
-            .map(|v| {
-                json!({
-                    "stateMachineVersionArn": format!("{}:{}", v.state_machine_arn, v.version),
-                    "creationDate": v.creation_date.timestamp(),
-                })
-            })
-            .collect();
-        let (page, token) = paginate(&items, next_token, max_results);
-        let mut resp = json!({ "stateMachineVersions": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn create_state_machine_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let name = body["name"]
-            .as_str()
-            .ok_or_else(|| missing("name"))?
-            .to_string();
-        validate_name(&name)?;
-        let routing_cfg = body["routingConfiguration"]
-            .as_array()
-            .ok_or_else(|| missing("routingConfiguration"))?;
-        let routes = parse_routing_configuration(routing_cfg)?;
-        let parent_arn = routes[0]
-            .state_machine_version_arn
-            .rsplit_once(':')
-            .map(|(parent, _)| parent.to_string())
-            .unwrap_or_default();
-        let alias_arn = format!("{parent_arn}:{name}");
-        let now = chrono::Utc::now();
-        let alias = crate::state::StateMachineAlias {
-            name,
-            arn: alias_arn.clone(),
-            description: body["description"].as_str().unwrap_or("").to_string(),
-            routing_configuration: routes,
-            creation_date: now,
-            update_date: now,
-        };
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        state.state_machine_aliases.insert(alias_arn.clone(), alias);
-        Ok(AwsResponse::ok_json(json!({
-            "stateMachineAliasArn": alias_arn,
-            "creationDate": now.timestamp(),
-        })))
-    }
-
-    fn delete_state_machine_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineAliasArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineAliasArn"))?
-            .to_string();
-        validate_arn_length("stateMachineAliasArn", &arn, 256)?;
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        state.state_machine_aliases.remove(&arn);
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    fn describe_state_machine_alias(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineAliasArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineAliasArn"))?
-            .to_string();
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let alias = state
-            .state_machine_aliases
-            .get(&arn)
-            .ok_or_else(|| resource_not_found(&arn))?;
-        Ok(AwsResponse::ok_json(state_machine_alias_to_json(alias)))
-    }
-
-    fn list_state_machine_aliases(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let parent = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?
-            .to_string();
-        validate_arn_length("stateMachineArn", &parent, 256)?;
-        let raw_max_results = body["maxResults"].as_i64();
-        if let Some(mr) = raw_max_results {
-            validate_max_results(mr)?;
-        }
-        let next_token = body["nextToken"].as_str();
-        if let Some(t) = next_token {
-            validate_page_token(t)?;
-        }
-        let max_results = match raw_max_results.unwrap_or(0) {
-            0 => 100,
-            n => n as usize,
-        };
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        // Anchor the prefix on the alias separator so a state machine
-        // named `foo` doesn't pull in aliases for `foobar`.
-        let parent_prefix = format!("{parent}:");
-        let mut aliases: Vec<&crate::state::StateMachineAlias> = state
-            .state_machine_aliases
-            .values()
-            .filter(|a| a.arn.starts_with(&parent_prefix))
-            .collect();
-        aliases.sort_by(|a, b| a.name.cmp(&b.name));
-        let items: Vec<Value> = aliases
-            .iter()
-            .map(|a| {
-                json!({
-                    "stateMachineAliasArn": a.arn,
-                    "creationDate": a.creation_date.timestamp(),
-                })
-            })
-            .collect();
-        let (page, token) = paginate(&items, next_token, max_results);
-        let mut resp = json!({ "stateMachineAliases": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn update_state_machine_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["stateMachineAliasArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineAliasArn"))?
-            .to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let alias = state
-            .state_machine_aliases
-            .get_mut(&arn)
-            .ok_or_else(|| resource_not_found(&arn))?;
-        if let Some(d) = body["description"].as_str() {
-            alias.description = d.to_string();
-        }
-        if let Some(routes) = body["routingConfiguration"].as_array() {
-            alias.routing_configuration = parse_routing_configuration(routes)?;
-        }
-        alias.update_date = chrono::Utc::now();
-        Ok(AwsResponse::ok_json(json!({
-            "updateDate": alias.update_date.timestamp(),
-        })))
-    }
-
-    // ─── Map runs ───────────────────────────────────────────────────────
-
-    fn describe_map_run(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["mapRunArn"]
-            .as_str()
-            .ok_or_else(|| missing("mapRunArn"))?
-            .to_string();
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let mr = state
-            .map_runs
-            .get(&arn)
-            .ok_or_else(|| resource_not_found(&arn))?;
-        Ok(AwsResponse::ok_json(map_run_to_json(mr)))
-    }
-
-    fn list_map_runs(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        // `executionArn` is required + has @length 1..=256.
-        let exec_arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?
-            .to_string();
-        validate_arn_length("executionArn", &exec_arn, 256)?;
-        let raw_max_results = body["maxResults"].as_i64();
-        if let Some(mr) = raw_max_results {
-            validate_max_results(mr)?;
-        }
-        let next_token = body["nextToken"].as_str();
-        if let Some(t) = next_token {
-            validate_page_token(t)?;
-        }
-        let max_results = match raw_max_results.unwrap_or(0) {
-            0 => 100,
-            n => n as usize,
-        };
-        let accounts = self.state.read();
-        let empty = crate::state::StepFunctionsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let mut runs: Vec<&crate::state::MapRun> = state
-            .map_runs
-            .values()
-            .filter(|r| r.execution_arn == exec_arn)
-            .collect();
-        // Stable order so pagination is deterministic.
-        runs.sort_by_key(|r| r.start_date);
-        let items: Vec<Value> = runs
-            .iter()
-            .map(|r| {
-                json!({
-                    "mapRunArn": r.map_run_arn,
-                    "executionArn": r.execution_arn,
-                    "stateMachineArn": "",
-                    "startDate": r.start_date.timestamp(),
-                    "stopDate": r.stop_date.map(|d| d.timestamp()),
-                })
-            })
-            .collect();
-        let (page, token) = paginate(&items, next_token, max_results);
-        let mut resp = json!({ "mapRuns": page });
-        if let Some(t) = token {
-            resp["nextToken"] = json!(t);
-        }
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn update_map_run(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["mapRunArn"]
-            .as_str()
-            .ok_or_else(|| missing("mapRunArn"))?
-            .to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let mr = state
-            .map_runs
-            .get_mut(&arn)
-            .ok_or_else(|| resource_not_found(&arn))?;
-        if let Some(c) = body["maxConcurrency"].as_i64() {
-            mr.max_concurrency = c as i32;
-        }
-        if let Some(p) = body["toleratedFailurePercentage"].as_f64() {
-            mr.tolerated_failure_percentage = p;
-        }
-        if let Some(c) = body["toleratedFailureCount"].as_i64() {
-            mr.tolerated_failure_count = c;
-        }
-        Ok(AwsResponse::ok_json(json!({})))
-    }
-
-    // ─── Execution lifecycle extras ─────────────────────────────────────
-
-    fn redrive_execution(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let arn = body["executionArn"]
-            .as_str()
-            .ok_or_else(|| missing("executionArn"))?
-            .to_string();
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let exec = state.executions.get_mut(&arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ExecutionDoesNotExist",
-                format!("Execution does not exist: {arn}"),
-            )
-        })?;
-        exec.status = crate::state::ExecutionStatus::Running;
-        exec.stop_date = None;
-        Ok(AwsResponse::ok_json(json!({
-            "redriveDate": chrono::Utc::now().timestamp(),
-        })))
-    }
-
-    async fn start_sync_execution(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let sm_arn = body["stateMachineArn"]
-            .as_str()
-            .ok_or_else(|| missing("stateMachineArn"))?
-            .to_string();
-        let input = body["input"].as_str().unwrap_or("{}").to_string();
-        if serde_json::from_str::<serde_json::Value>(&input).is_err() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidExecutionInput",
-                "Execution input is not valid JSON.",
-            ));
-        }
-        let (exec_arn, definition, logging_config) = {
-            let mut accounts = self.state.write();
-            let state = accounts.get_or_create(&req.account_id);
-            let sm = state
-                .state_machines
-                .get(&sm_arn)
-                .ok_or_else(|| state_machine_not_found(&sm_arn))?;
-            if sm.machine_type != crate::state::StateMachineType::Express {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "StateMachineTypeNotSupported",
-                    "StartSyncExecution is only supported for EXPRESS state machines.",
-                ));
-            }
-            let now = chrono::Utc::now();
-            let exec_name = format!("sync-{}", now.timestamp_millis());
-            let exec_arn = format!(
-                "arn:aws:states:{}:{}:express:{}:{}",
-                state.region, state.account_id, sm.name, exec_name
-            );
-            let execution = Execution {
-                execution_arn: exec_arn.clone(),
-                state_machine_arn: sm_arn.clone(),
-                state_machine_name: sm.name.clone(),
-                name: exec_name.clone(),
-                status: ExecutionStatus::Running,
-                input: Some(input.clone()),
-                output: None,
-                start_date: now,
-                stop_date: None,
-                error: None,
-                cause: None,
-                history_events: vec![],
-                parent_execution_arn: None,
-                is_sync: true,
-                billed_duration_ms: None,
-                billed_memory_mb: None,
-            };
-            state.executions.insert(exec_arn.clone(), execution);
-            (
-                exec_arn,
-                sm.definition.clone(),
-                sm.logging_configuration.clone(),
-            )
-        };
-
-        interpreter::execute_state_machine(
-            self.state.clone(),
-            exec_arn.clone(),
-            definition,
-            Some(input),
-            self.delivery.clone(),
-            self.dynamodb_state.clone(),
-            self.registry.clone(),
-            logging_config,
-        )
-        .await;
-
-        // Persist billing details on the stored execution so introspection
-        // endpoints can replay the same numbers later.
-        {
-            let mut accounts = self.state.write();
-            if let Some(state) = accounts.get_mut(&req.account_id) {
-                if let Some(exec) = state.executions.get_mut(&exec_arn) {
-                    let duration_ms = exec
-                        .stop_date
-                        .map_or(0, |stop| (stop - exec.start_date).num_milliseconds())
-                        .max(0);
-                    exec.billed_duration_ms = Some(duration_ms);
-                    exec.billed_memory_mb = Some(64);
-                }
-            }
-        }
-
-        let accounts = self.state.read();
-        let state = accounts.get(&req.account_id).unwrap();
-        let exec = state
-            .executions
-            .get(&exec_arn)
-            .ok_or_else(|| execution_not_found(&exec_arn))?;
-
-        let mut resp = json!({
-            "executionArn": exec.execution_arn,
-            "stateMachineArn": exec.state_machine_arn,
-            "name": exec.name,
-            "startDate": exec.start_date.timestamp(),
-            "stopDate": exec.stop_date.map(|d| d.timestamp()),
-            "status": exec.status.as_str(),
-            "input": exec.input.as_deref().unwrap_or("{}"),
-        });
-
-        if let Some(ref output) = exec.output {
-            resp["output"] = json!(output);
-        }
-        if let Some(ref error) = exec.error {
-            resp["error"] = json!(error);
-        }
-        if let Some(ref cause) = exec.cause {
-            resp["cause"] = json!(cause);
-        }
-
-        let duration_ms = exec
-            .stop_date
-            .map_or(0, |stop| (stop - exec.start_date).num_milliseconds());
-        resp["billingDetails"] = json!({
-            "billedMemoryUsedInMB": 64,
-            "billedDurationInMilliseconds": duration_ms.max(0),
-        });
-
-        Ok(AwsResponse::ok_json(resp))
-    }
-
-    fn test_state(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let definition = body["definition"]
-            .as_str()
-            .ok_or_else(|| missing("definition"))?;
-        validate_definition(definition)?;
-        let _role_arn = body["roleArn"].as_str().ok_or_else(|| missing("roleArn"))?;
-        let input = body["input"].as_str().unwrap_or("{}").to_string();
-        // Echo input back as output. Real Step Functions actually
-        // simulates the state; our emulator reports SUCCEEDED so callers
-        // can wire the integration test scaffolding.
-        Ok(AwsResponse::ok_json(json!({
-            "output": input,
-            "status": "SUCCEEDED",
-            "nextState": "End",
-        })))
-    }
-
-    fn validate_state_machine_definition(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let definition = body["definition"]
-            .as_str()
-            .ok_or_else(|| missing("definition"))?;
-        if definition.is_empty() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "definition must be 1..=1048576 characters",
-            ));
-        }
-        if let Some(mr) = body["maxResults"].as_i64() {
-            if !(0..=100).contains(&mr) {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("maxResults '{mr}' is outside 0..=100"),
-                ));
-            }
-        }
-        if let Some(sev) = body["severity"].as_str() {
-            if !matches!(sev, "ERROR" | "WARNING") {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("severity '{sev}' must be ERROR or WARNING"),
-                ));
-            }
-        }
-        if let Some(ty) = body["type"].as_str() {
-            if !matches!(ty, "STANDARD" | "EXPRESS") {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("type '{ty}' must be STANDARD or EXPRESS"),
-                ));
-            }
-        }
-        match validate_definition(definition) {
-            Ok(()) => Ok(AwsResponse::ok_json(json!({
-                "result": "OK",
-                "diagnostics": [],
-            }))),
-            Err(e) => Ok(AwsResponse::ok_json(json!({
-                "result": "FAIL",
-                "diagnostics": [{
-                    "severity": "ERROR",
-                    "code": "INVALID_DEFINITION",
-                    "message": e.to_string(),
-                }],
-            }))),
-        }
     }
 }
